refactor: 공통 컴포넌트 폴더명 변경
