+++ src/component/@common/index.js
@@ -1,0 +1,3 @@
+export { default as Button } from 'component/@common/Button/Button';
+export { default as GridLayout } from 'component/@common/GridLayout/GridLayout';
+export { default as Spinner } from 'component/@common/Spinner/Spinner';
